@@ -11,41 +11,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.8",
-    "@angular/cdk": "^8.2.2",
-    "@angular/common": "~8.2.8",
-    "@angular/compiler": "~8.2.8",
-    "@angular/core": "~8.2.8",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "~8.2.8",
-    "@angular/material": "^8.2.2",
-    "@angular/platform-browser": "~8.2.8",
-    "@angular/platform-browser-dynamic": "~8.2.8",
-    "@angular/router": "~8.2.8",
-    "hammerjs": "^2.0.8",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "@angular/animations": "~11.0.2",
+    "@angular/cdk": "^11.0.1",
+    "@angular/common": "~11.0.2",
+    "@angular/compiler": "~11.0.2",
+    "@angular/core": "~11.0.2",
+    "@angular/flex-layout": "^11.0.0-beta.33",
+    "@angular/forms": "~11.0.2",
+    "@angular/material": "^11.0.1",
+    "@angular/platform-browser": "~11.0.2",
+    "@angular/platform-browser-dynamic": "~11.0.2",
+    "@angular/router": "~11.0.2",
+    "rxjs": "~6.6.3",
+    "tslib": "^2.0.3",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.6",
-    "@angular/cli": "~8.3.6",
-    "@angular/compiler-cli": "~8.2.8",
-    "@angular/language-service": "~8.2.8",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "@angular-devkit/build-angular": "~0.1100.2",
+    "@angular/cli": "~11.0.2",
+    "@angular/compiler-cli": "~11.0.2",
+    "@angular/language-service": "~11.0.2",
+    "@types/node": "^14.14.10",
+    "@types/jasmine": "~3.6.2",
+    "@types/jasminewd2": "~2.0.8",
+    "codelyzer": "^6.0.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~6.0.0",
+    "karma": "~5.2.3",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "~7.0.0",
+    "ts-node": "~9.0.0",
+    "tslint": "~6.1.3",
+    "typescript": "~4.0.5"
   }
 }
